Use new system for cross-platform package installation

+++ setup/30_neovim/setup.bash
@@ -1,10 +1,9 @@
 setup() {
-	install_pacman neovim
-	install_dnf neovim
+	install_pkg neovim || return 1
 
 	run ln -s "$(command -v nvim)" "$HOME/.local/bin/vim"
 
-	run mv ~/.config/nvim{,.bak}
+	[[ -f ~/.config/nvim ]] && run mv ~/.config/nvim{,.bak}
 	run git clone https://github.com/LazyVim/starter ~/.config/nvim
 	run rm -rf ~/.config/nvim/.git
 }
@@ -13,6 +12,7 @@ config() {
 	[[ -z "$DOTFILES_PATH" ]] && fatal 'variable DOTFILES_PATH needs to be set!'
 
 	run cp -R ~/.config/nvim ~/.config/nvim.old
-	run cp -R "$DOTFILES_PATH"/.config/nvim/* ~/.config/nvim/
-	run cp -R "$DOTFILES_PATH"/.config/nvim/.* ~/.config/nvim/
+	for file in "$DOTFILES_PATH"/.config/nvim/.* "$DOTFILES_PATH"/*; do
+		run cp -R "$file" ~/.config/nvim/
+	done
 }
